requirements.txt
biopython==1.81
streamlit==1.25.0
matplotlib==3.7.2